Use common package to set up ginkgo Context
Instead of explicitly calling a function that sets up the common tests
we can leverage the package that we anyway import. This removes the need
to explicitly call the `common.Run` function in all the suites. Just
importing the package will set up the tests.
This has the downside of being less explicit. We have to make this clear
somehow that these tests are set up if we go this route.