Add common tests checking if nodes are in desired ranges (#19)
* Add tests to verify expected node counts

* Account for providers that only have replicas

* Change function names

* Duplicate node tests

cluster-azure has a different schema than the rest of the providers,
which means that the values cannot be parsed the same way. To
temporarily fix this while the schemas get aligned, the same test needs
to be duplicated in the providers so we can use a specific values struct
just for capz. When the schema has been aligned with the rest of the
providers we can put the tests back in the common package.

* Change label matcher for CAPZ

* bump clustertest to 0.0.14